Extract Pro WithTax into Taxable concern
Replacing wrapped Stripe objects with new Pro Plan which includes
Taxable.
