+++ app/models/alaveteli_pro/plan.rb
@@ -1,0 +1,8 @@
+##
+# A wrapper for a Stripe::Plan
+#
+class AlaveteliPro::Plan < SimpleDelegator
+  include Taxable
+
+  tax :amount
+end
